Ask to create the base build directory if it does not exist yet.
