+++ pprof/generate_config.py
@@ -1,5 +1,6 @@
 from plumbum import cli
 from pprof.driver import PollyProfiling
+from pprof.utils.user_interface import query_yes_no
 import sys
 import os.path
 
@@ -33,37 +34,3 @@ class PprofGenConfig(cli.Application):
                 outfile.write("\n\n")
 
         print("Configuration file has been written to " + self._outpath)
-
-
-# Taken from the following recipe: http://code.activestate.com/recipes/577058/
-def query_yes_no(question, default="yes"):
-    """Ask a yes/no question via raw_input() and return their answer.
-
-    "question" is a string that is presented to the user.
-    "default" is the presumed answer if the user just hits <Enter>.
-        It must be "yes" (the default), "no" or None (meaning
-        an answer is required of the user).
-
-    The "answer" return value is True for "yes" or False for "no".
-    """
-    valid = {"yes": True, "y": True, "ye": True,
-             "no": False, "n": False}
-    if default is None:
-        prompt = " [y/n] "
-    elif default == "yes":
-        prompt = " [Y/n] "
-    elif default == "no":
-        prompt = " [y/N] "
-    else:
-        raise ValueError("invalid default answer: '%s'" % default)
-
-    while True:
-        sys.stdout.write(question + prompt)
-        choice = raw_input().lower()
-        if default is not None and choice == '':
-            return valid[default]
-        elif choice in valid:
-            return valid[choice]
-        else:
-            sys.stdout.write("Please respond with 'yes' or 'no' "
-                             "(or 'y' or 'n').\n")
